Workaround for "empty" output error for gnuradio metapackage.
Without a script or files defined, conda-build emits a warning and then
does a check that Python is not in both the build and host environments.
Unfortunately, due to ninja, python is in the build environment now, so
conda-build raises an exception. This puts in a dummy script for the
gnuradio metapackage output to avoid triggering that check.