Add patch to fix exported gnuradio-runtime CMake target.
The current exported target contains absolute include paths, including
one to numpy in the build directory that happens because of
cross-compilation. The build directory path is not detected/replaced
through the typical prefix path processing, so it results in errors when
the target is used (e.g. by GR OOT modules). This cleans up the target
so that it doesn't have any absolute paths.