Hack the solver to get the output dependencies working.
We're in the middle of a migration for zstd, and elfutils hasn't been
migrated yet, but most of the outputs don't know that they need to use
the older zstd to accomodate elfutils because apparently it doesn't take
into account the subpackage dependencies (i.e. gnuradio-core) before
freezing the versions required of the output's dependencies. This adds
elfutils to the host environment (ignore its run_exports) so that it
will choose compatible dependencies.