Rename gnuradio-soapy to gnuradio-soapy-intree to avoid conflict.
We can't archive the separate gnuradio-soapy feedstock yet because it's
still needed to support GR 3.8, and conda-forge doesn't allow two
feedstocks to upload packages to the same name. So we'll call this one
gnuradio-soapy-intree for now, and we can rename it back to
gnuradio-soapy once the other feedstock can be archived.