Modify jinja to fix rendering for cf-scripts pinning update.
When cf-scripts (regro-cf-autotick-bot) parses the meta.yaml to
determine if the global pinning needs to be updated, it uses stub
versions of the jinja functions instead of the actual versions. In this
case, `pin_subpackage` returns a dictionary instead of a string, so the
string parsing for the build string of gnuradio-build-deps causes an
exception. Adding logic to ensure that the build string is empty when
`pin_subpackage` returns a non-string fixes this.