Remove gnuradio-build-deps host dep on gnuradio-core
It's a metapackage anyway, and the pinned subpackage doesn't put its
run_exports into the run deps of the metapackage, so removing it should
do nothing. But currently having it there triggers a solver bug where it
can't find a compatible version of libxml2 to use. Hopefully doing this
avoids that problem, and it's not still an issue from the run deps.